requirements.txt
fastapi[standard]==0.115.0
Faker==30.4.0
python-jose[cryptography]==3.3.0